Detect the player’s tile position in Minecraft Pi Edition and send data to an Arduino sketch called pumpkinPi.ino to light RGB NeoPixels.
Built and coded on the floor of Minecon 2013, in Orlando, FL. Credit
for build goes to Theodore Wahrburg, Eli Rutan, and Mac Rutan.

http://familab.org/blog/2013/11/familab-hosts-the-minecon2013-maker-vill
age/

Overview:

The “mineLight.py” script is the python component used the in Minecraft
Pi Edition API to detect and send block identification data via serial
communication to an adruino that controls LED neopixels.

Links to the other software components, including the pumpkinPi.ino
Arduino sketch are included here:

https://github.com/familab

See the read me for detailed instructions.